requirements.txt
flask
tensorflow
numpy
pillow
requests

